+++ source/MapFactory.ts
@@ -6,20 +6,23 @@ class MapFactory {
     }
 
     public create(width: number, height: number): EditorMap {
+        const layers = [
+            LayersHelper.create('text', this.localizer['layer_type_text']),
+            LayersHelper.create('grid', this.localizer['layer_type_grid']),
+            LayersHelper.create('terrain', this.localizer['layer_type_terrain'])
+        ];
+
         return {
             zoom: 2,
             position: { x: .5, y: .5 },
+            activeLayer: layers[0].id,
             data: {
                 id: Utilities.generateId('map'),
                 columns: width,
                 rows: height,
                 pixelsPerCell: 100,
                 objects: {},
-                layers: [
-                    LayersHelper.create('text', this.localizer['layer_type_text']),
-                    LayersHelper.create('grid', this.localizer['layer_type_grid']),
-                    LayersHelper.create('terrain', this.localizer['layer_type_terrain'])
-                ]
+                layers
             }
         };
     }